pyproject.toml
[tool.poetry]
name = "julius"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "20.7"
anthropic = "0.39.0"
gspread = "6.1.2"
google-auth = "2.35.0"
google-auth-oauthlib = "1.2.1"
google-auth-httplib2 = "0.2.0"
python-dotenv = "1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
